Update article controller tests
